requirements.txt
notebook
ipywidgets
numpy
pandas
Pillow